perf: avoid copies in AttributeHandler
